Committing drafts of scripts for initial MC sim of RTS-GMLC example
run_all_MC_sim - currently reads in all zone 1 csvs and inputs into DF
run_all_seq - currently runs simple modifications of one input file and stores results

+++ run_all_seq.py
@@ -50,7 +50,11 @@ def modify_file(path):
 
 
 def copy_directory(index):
-    shutil.copytree(dir_path, new_path)
+    if os.path.exists(new_path):
+        shutil.rmtree(new_path)
+        shutil.copytree(dir_path, new_path)
+    else:
+        shutil.copytree(dir_path, new_path)
 
 
 def run(i):
